Fixed #38, Gevent now supports Python 3.3 as single source.
The tests are all passing now. But some of the Python 3 stdlib
interfaces are not implemented yet. And, there are still lots of
ResourceWarnings for unclosed files or sockets in Python 3.

These modules are fixed with explict PY3 code fragment branches:

    ares, backdoor, core, fileobject, hub, os, pywsgi, queue,
    server, socket, ssl, subprocess

The Six library is partly introduced as a part of hub.py (while
for tests it is greentest/six.py). These symbols are added:

    text_type, binary_type, xrange, b, u, reraise

Travis can test under Python 3.3 for now, and ignoring tests
depending on web.py which doesn't support Python 3 yet.